+++ doc/CHANGELOG.md
@@ -6,7 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
-## Version 1.0.5 - tbd.
+## Version 1.0.5 - 2024-11-06
 
 ### Added